rename benchstat to summarize
