+++ go/summarize/testdata/trace-log.json
